requirements.txt
pip
Flask
requests
sqlalchemy
passlib
flask_cors
pyjwt
